requirements.txt
aliyun-python-sdk-core==2.13.36
aliyun-python-sdk-kms==2.16.1
amqp==2.6.1
asgiref==3.6.0
asn1crypto==1.5.1
async-timeout
backoff==2.2.1
billiard==3.6.4.0
blinker==1.6.2
Bottleneck
brotlipy==0.7.0
celery==4.4.7
cffi
charset-normalizer
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
coreapi==2.3.3
coreschema==0.0.4
crcmod==1.7
cryptography
Deprecated==1.2.13
Django==4.2
django-cors-headers==3.14.0
django-redis==5.2.0
django-requestlogs==0.7.1
django-rest-swagger==2.2.0
django-snowflake==4.2b1
djangorestframework==3.14.0
drf-extensions==0.7.1
drf-yasg==1.21.5
filelock==3.12.0
Flask==2.3.1
googleapis-common-protos==1.56.2
gunicorn==20.1.0
h11==0.14.0
idna
importlib-metadata==6.0.1
inflection==0.5.1
install==1.3.5
itsdangerous==2.1.2
itypes==1.2.0
jaeger-client==4.8.0
Jinja2==3.1.2
jmespath==0.10.0
kombu==4.6.11
MarkupSafe==2.1.2
numexpr
openapi-codec==1.3.2
opentracing==2.4.0
oscrypto==1.3.0
oss2==2.18.0
packaging>=23.2,<24.0
pandas==1.5.3
prompt-toolkit==3.0.38
protobuf==3.20.3
pycparser
pycryptodome==3.17
pycryptodomex==3.17
PyJWT==2.7.0
PyMySQL==1.0.3
pyOpenSSL
PySocks
python-alipay-sdk==3.2.0
python-dateutil
pytz==2023.3
redis==4.5.4
redis-py-cluster
requests
ruamel.yaml==0.17.26
ruamel.yaml.clib==0.2.7
simplejson==3.19.1
six
snowflake-connector-python==3.0.4
sortedcontainers==2.4.0
sqlparse
threadloop==1.0.2
thrift==0.16.0
tornado==6.3.2
typing_extensions
tzdata==2023.3
uritemplate==4.1.1
urllib3
uvicorn==0.22.0
vine==1.3.0
wcwidth==0.2.6
wechatpayv3==1.2.37
Werkzeug==2.3.0
wrapt
zipp==3.15.0
zope.event==4.6
zope.interface==6.0
django-ratelimit
requests-futures
pika
requests_toolbelt
gevent
websocket-client
openai
PyPDF2==3.0.1
python-docx-1
# embeded-chain requirements
beautifulsoup4==4.12.2
certifi==2023.11.17
chromadb==0.4.21
fastapi==0.106.0
fastapi_poe==0.0.25
flask_sqlalchemy==3.1.1
GitPython==3.1.40
gradio==4.12.0
langchain==0.0.352
llama_hub==0.0.64
llama_index==0.9.21
mock==5.1.0
modal==0.56.4376
Pillow==10.1.0
psycopg==3.1.16
pydantic==2.5.3
pymilvus==2.3.4
pytest==7.4.0
python-dotenv==1.0.0
PyYAML==6.0.1
Requests==2.31.0
responses==0.24.1
rich==13.7.0
schema==0.7.5
sentence_transformers==2.2.2
setuptools==68.2.2
slack_sdk==3.26.1
SQLAlchemy==2.0.23
streamlit==1.29.0
tqdm==4.66.1
twilio==8.11.0
wikipedia==1.4.0
yt_dlp==2023.11.16
pypdf==3.17.4
opensearch-py==2.4.2
docx2txt